Implements feedback
