+++ server/pom.xml
@@ -246,6 +246,7 @@
         <dependency>
             <groupId>org.json</groupId>
             <artifactId>json</artifactId>
+            <version>20090211</version>
         </dependency>
     </dependencies>
 
@@ -367,17 +368,6 @@
                                 <Kaazing-Product>Community.Gateway</Kaazing-Product>
                             </manifestEntries>
                         </archive>
-                    </configuration>
-                </plugin>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-surefire-plugin</artifactId>
-                    <configuration>
-                        <additionalClasspathElements>
-                            <additionalClasspathElement>${basedir}/src/test/resources/gateway.server-5.0.0.8.jar</additionalClasspathElement>
-                        </additionalClasspathElements>
-                        <argLine>${surefireArgLine}</argLine>
-                        <skipTests>${skipTests}</skipTests>
                     </configuration>
                 </plugin>
             </plugins>
