+++ update.check/src/main/java/org/kaazing/gateway/update/check/UpdateCheckLoggingListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kaazing.gateway.server.update.check;
+package org.kaazing.gateway.update.check;
 
 import static java.lang.String.format;
 import static org.kaazing.gateway.server.impl.VersionUtils.getGatewayProductTitle;
@@ -23,7 +23,7 @@ import org.slf4j.LoggerFactory;
 
 /**
  * On notifications of new versions, this class logs the information
- * 
+ *
  */
 class UpdateCheckLoggingListener implements UpdateCheckListener {
 
@@ -34,10 +34,5 @@ class UpdateCheckLoggingListener implements UpdateCheckListener {
         logger.info(format(
                 "Update Check: New release available for download: %s %s (you are currently running %s)",
                 getGatewayProductTitle(), latestGatewayVersion, currentVersion));
-    }
-
-    @Override
-    public void setUpdateCheck(UpdateCheck updateCheck) {
-        // NOOP, don't care about the service
     }
 }
