+++ update.check/src/main/resources/META-INF/services/org.kaazing.gateway.server.GatewayObserverFactorySpi
@@ -1,0 +1,1 @@
+org.kaazing.gateway.update.check.UpdateCheckGatewayObserver
