+++ update.check/src/test/scripts/org/kaazing/gateway/update/check/testUpdateCheckTaskWithFailedRequestsResponseCode.rpt
